Add methods for line wiping and screen clearing.
In this context, line wiping refers to temporarily removing the contents
of the current line from the screen, but not otherwise changing the line
editing state.  Its use case is temporarily removing the prompt to allow
other output to be written, with an intent to immediately restore it
once that output is complete.