Make linenoiseState a class with methods, various other adjustments (#34)
* Reduce variable scope

* Make most linenoise functions into methods on a class.

* Support previous API style

Implement a way to allow the previous API as seen in the old example.cpp to
keep working.

* Add a mutex to protect RefreshLine if multiple threads need to call it.

* Adjust how history and its up/down arrow browsing are handled.

* We don't want to hang if we can't read anything from stdin

* Add WipeLine and ClearScreen methods.

* See if we can avoid using strcasecmp

* Correct behavior for sending commands

Fix from Chris McGregor (f4alt <christopherjmacgregor@gmail.com>:
Old linenoise behavior would send commands on newline. For subprocess use, like
we do with nirt in MGED, this is very important as the alternative means
collecting commands like "Command_0\nCommnand_1.." and expecting the process to
parse the newlines.

* Misc. fixes and adjustments

* Compare against the correct string

* Use inline

* Restore inlines to allow multiple includes

Also add a compilation test case to examples to verify we can include more than
once and create an executable.

* Add an l_ prefix to all class member variables.

* Stray static global mess with history rename.  Add some comments.

* Eliminate a couple more globals

* Back out history changes

* Use suffix character for class variables

* Remove unnecessary prefixes from internal class method names

* Clean up class API slightly.

* Unused define

* Move multiinclude test files to subdir